refactor: move schema-analyze to schema/ and extract LintContext to lint/types
- Move src/wiki/schema-analyze.ts to src/schema/analyze.ts (git mv).
- Mirror test file: src/__tests__/wiki/schema-analyze.test.ts →
  src/__tests__/schema/analyze.test.ts.
- Update main.ts import path.
- Extract LintContext from lint-controller.ts to lint/types.ts so
  fix-runners.ts no longer depends on lint-controller.ts (breaks
  the latent cycle).
- lint-controller.ts re-exports LintContext for back-compat.
- 4-Gate clean; 728 tests passing.
