+++ src/schema/analyze.ts
